pyproject.toml
[tool.poetry]
name = "fastapi-jwt-middleware"
version = "1.0.0"
description = "Lightweight JSON web token authorisation middleware for FastAPI."
authors = ["Clark Sirl <click66@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.1"
cryptography = "^41.0.3"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
